// client-cargo-main (1)/client-cargo-main/src/stores/Trip/Trip.interface.ts
export interface ITripRequestData {
  pageSetting: { page: number; size: number };
  desiredDate?: { start?: number | null; end?: number | null };
  creationDate?: { start?: number | null; end?: number | null };
  page?: { sort: { sorted: boolean } };
  sortSetting?: { directionAsc: boolean };
  transportTypeEnum?: TransportTypeEnum;
}

export interface ITripService {
  calculateTripCost: (
    data: ITripCalculateRequest,
    callBack: (transportTypeId: string, tariffId: UUID) => {}
  ) => Promise<ITripTariff[]>;
  calculateCostForPersonalCarTrip: (data: ITripCalculateRequest) => Promise<ITripTariff[]>;
  calculateExternalPrices: (data: TTripRequestNew) => Promise<TExternalPrices[]>;

  saveTripRequest(data: TTripRequestNew): Promise<TripRequest | number>;

  editTripRequest(data: TTripRequestNew, requestId: string): Promise<number>;
  rateTripRequest(data: IRequestRating, requestId: string): Promise<number>;
  getTripRequestList(passengerId: string): Promise<TripRequest[]>;
  getTripRequestListNonTerminal(data: ITripRequestData): Promise<TripPaginationModel & { content: TripRequest[] }>;
  getTripRequestListTerminal(data: ITripRequestData): Promise<TripPaginationModel & { content: TripRequest[] }>;
  getTripRequest(requestId: string): Promise<TripRequest>;

  getAllPurposesByEmployee(orgId: string): Promise<TripPurpose[]>;
  getAllTariffsTaxi(): Promise<TTariffTaxi[]>;
  getAllTariffsPersonal(): Promise<TTariffPersonal[]>;
  getAllTariffsPublic(): Promise<TTariffPublic>;
  getTariffById(transTypeId: string, tariffId: UUID): Promise<TTariffPublic>;

  getAllCoopTrips(data: TripRequest): Promise<TTripCoop[]>;
  joinCoopTrip(sharedId: string, data: TTripRequestNew): Promise<TripRequest>;

  finishTripRequest(reqId: string): Promise<number>;

  saveFile(file: any, folder: UUID | string): Promise<SavedFileInfo>;
  saveConfirmSuburbTripFile(requestId: string | UUID, filesData: SavedFileInfo[]): Promise<number>;
  saveConfirmCardTripFile(requestId: string | UUID, filesData: SavedFileInfo[]): Promise<number>;
  saveTravelCardRequest(data: InnerCityTransportRequestInfo): Promise<number>;
  saveSuburbCompensation(data: InnerCityTransportRequestInfo): Promise<number>;
  saveCityTripCompensation(data: CityTripCompensation): Promise<number>;
  savePublicTripRequest(data: PublicTripCompensation): Promise<number>;
  loadCoopTripsSettings(organizationId: string): Promise<TCoopTripsSettings[]>;
}

export type ITripServiceResponse = Record<string, any>;

export interface ITripCalculateRequest {
  distance: number;
  time: number;
  tripDate: number;
  startPoint: Record<string, any>;
  organizationId: string;
  engineVolume?: number;
  waitingTime: number;
  intermediateWaitingTime: number;
}

export type Transport = TaxiClassEnum | TransportTypeEnum;

export enum TaxiClassEnum {
  ECONOMY = 'ECONOMY',
  COMFORT = 'COMFORT',
  PERSONAL = 'PERSONAL',
  BUSINESS = 'BUSINESS',
  TAXI = 'TAXI',
  CARSHARING = 'CARSHARING',
  BICYCLE = 'BICYCLE',
  WALK = 'WALK',
  PUBLIC = 'PUBLIC',
  SCOOTER = 'SCOOTER',
  YANDEX = 'YANDEX',
  CITYMOBIL = 'CITYMOBIL',
  UBER = 'UBER',

  VIP_BUS = 'VIP_BUS',
  SMALL_BUS = 'SMALL_BUS',
  MIDDLE_BUS = 'MIDDLE_BUS',
  LARGE_BUS = 'LARGE_BUS',
}
// ...
export type TTaxiClass = keyof typeof TaxiClassEnum;
// ...
export type TExternalPrices = t.TypeOf<typeof IOExternalPrices>;

export interface ITaxi {
  name: string;
  transportType: TransportTypeEnum;
  taxiClass?: TTaxiClass;
  waitingTime: number;
  disabled?: boolean;
  info?: string;
  type?: LIMIT_TYPE;
  contractorId?: string;
}

export interface ITripTariff {
  id: string;
  cost: number;
  transportType: { id: string; name: TransportTypeEnum };
  taxiClass?: TTaxiClass;
  contractorId?: string;

  bonusCost?: number | undefined;
}
// ...
export const IOTariffBase = t.intersection([
  t.type({
    rideCostPerKm: t.number,
  }),
  t.partial({
    name: t.string,
    regionId: t.string,
    // Сомнительное место. Паша должен исправить.
    organizationId: t.string,
    transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
  }),
]);

export const IOTariffTaxi = t.intersection([
  IOTariffBase,
  t.type({
    taxiClass: ioTypeFromEnum<TaxiClassEnum>('TaxiClassEnum', TaxiClassEnum),
    rideCostPerMin: t.number,
    waitCostPerMin: t.number,
    freeWaitingTime: t.number,
  }),
  t.partial({
    id: t.string,
    priceDetails: t.unknown,
    deptStruct: t.string,
    deptName: t.string,
    vehicleType: t.string,
    minRideCost: t.number,
    savingsDeviationPct: t.number,
    distanceDeviationKm: t.number,
    timeDeviationMin: t.number,
    maxCapacity: t.number,
    minCancelTimeMin: t.number,
    tollRoads: t.boolean,
    historicalTraffic: t.boolean,
    minServiceTime: t.number,
    tariffId: t.string,
    active: t.boolean,
    contractorId: t.string,
  }),
]);

export const IOTariffPriceDetails = t.partial({
  rewardForPassenger: t.number,
  season: t.partial({
    start: t.string,
    end: t.string,
  }),
});

export const IOTariffPersonal = t.intersection([
  IOTariffBase,
  t.type({
    seasonStart: t.string,
    seasonEnd: t.string,
  }),
  t.partial({
    id: t.string,
    priceDetails: IOTariffPriceDetails,
    rewardForPassenger: t.number,
  }),
]);

export type TTariffPriceDetails = t.TypeOf<typeof IOTariffPriceDetails>;

export type TTariffBase = t.TypeOf<typeof IOTariffBase>;

export type TTariffTaxi = t.TypeOf<typeof IOTariffTaxi>;

export type TTariffPersonal = t.TypeOf<typeof IOTariffPersonal>;

export type TTarifUnited = TTariffTaxi | TTariffPersonal;

export const IOTariffPublicDetails = t.type({
  description: t.string,
});

export const IOTariffPublic = t.intersection([
  t.type({
    id: t.string,
    humanReadableId: t.string,
    active: t.boolean,
    regionId: t.string,
    transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
    serviceType: t.string,
    organizationId: t.string,
    metroTicketCost: tt.money,
    tramTicketCost: tt.money,
    trolleybusTicketCost: tt.money,
    busTicketCost: tt.money,
    metroAvailability: t.boolean,
    tramAvailability: t.boolean,
    trolleybusAvailability: t.boolean,
    busAvailability: t.boolean,
    cityLocalTrainAvailability: t.boolean,
    cityLocalTrainCost: tt.money,
    // priceDetails: Proxy
    travelCardAllCityTransportAvailability: t.boolean,
    travelCardAllCityTransportCost: tt.money,
    travelCardBusAvailability: t.boolean,
    travelCardBusCost: tt.money,
    travelCardLocalTrainAvailability: t.boolean,
    travelCardLocalTrainCost: tt.money,
    travelCardMetroAvailability: t.boolean,
    travelCardMetroCost: tt.money,
    travelCardTramAvailability: t.boolean,
    travelCardTramCost: tt.money,
    travelCardTrolleybusAvailability: t.boolean,
    travelCardTrolleybusCost: tt.money,
  }),
  t.partial({
    compensationType: ioTypeFromEnum<TransportCompensations>('TransportCompensations', TransportCompensations),
    // deprecated
    region: t.string,
  }),
]);

export type TTariffPublic = t.TypeOf<typeof IOTariffPublic>;
// ...
export const TripPurpose = t.intersection([
  t.type({
    // id: tt.uuid,
    id: t.string,
    label: t.string,
  }),
  t.partial({
    tripPurposeAttributes: t.array(IOTripPurposeAttributes),
    tripPurposeDates: t.array(IOTripPurposeDates),
    tripPurposeTimes: t.array(IOTripPurposeTimes),
    tripPurposeWeekdays: t.array(IOTripPurposeWeekdays),
  }),
]);

export type TripPurpose = t.TypeOf<typeof TripPurpose>;

export const IODeclineReason = t.type({
  reason: t.string,
  code: t.number,
});

export type IDeclineReason = t.TypeOf<typeof IODeclineReason>;

// Ratings

export interface IRequestRating {
  rating?: number;
  advantages?: string[];
  drawbacks?: string[];
  ratingComment?: string;
}

export const IORequestRating = t.partial({
  rating: t.number,
  advantages: t.array(t.string),
  drawbacks: t.array(t.string),
  ratingComment: t.string,
});

export type TRequestRating = t.TypeOf<typeof IORequestRating>;

// Trip Requests

export const TripRequest = t.intersection([
  IOHumanReadable,
  t.type({
    id: t.string,
    author: Employee,
    passenger: Employee,
    purpose: TripPurpose,
    passengerCount: t.number,
    desiredDate: t.number,
    creationTime: t.union([t.string, t.number]),
    approvedBy: Employee,
    coopTrip: t.boolean,
  }),
  t.partial({
    commentForDriver: t.string,
    humanReadableId: t.string,
    restOfLimit: t.number,
    transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
    status: TTripRequestStatuses,
    approvalState: ApprovalStateStatuses,
    expected: RequestRoute,
    taxiClass: ioTypeFromEnum<TaxiClassEnum>('TaxiClassEnum', TaxiClassEnum),
    tariffId: t.string,
    requestRating: IORequestRating,
    requestOptions: t.array(t.string),
    passengerLimitRemainder: t.array(LimitSharing),
    magentaOrderId: t.number,
    sharedRideId: t.number,
    occupiedPlacesCount: t.number,
    personalCar: IOPersonalCar,
    compensationType: ioTypeFromEnum<TransportCompensations>('TransportCompensations', TransportCompensations),
    timeZone: t.string,
  }),
]);

export type TripRequest = t.TypeOf<typeof TripRequest>;

export type TTripRequestNew = Pick<TripRequest, 'author' | 'passenger' | 'transportType' | 'expected' | 'purpose'> &
  Partial<TripRequest>;

// Suitable Trips

export const IOTripCancelReason = t.type({
  reason: t.string,
  code: t.number,
});

export type TripCancelReason = t.TypeOf<typeof IOTripCancelReason>;

export const IOTripStops = t.intersection([
  t.type({
    orderId: t.number,
    address: t.string,
    latitude: t.number,
    longitude: t.number,
    startTime: t.union([t.number, t.string]),
    endTime: t.union([t.number, t.string]),
    eventType: t.string,
  }),
  t.partial({
    state: t.string,
  }),
]);
// ...
export const IOTripCoop = t.type({
  id: t.string,
  passengers: t.number,
  employeePassengers: t.array(Employee),
  newOrdersIds: t.array(t.string),
  stops: t.array(IOTripStops),
  kpi: IOTripKpi,
});

export type TTripStops = t.TypeOf<typeof IOTripStops>;
export type TTripFromCoopStops = t.TypeOf<typeof IOTripFromCoopStops>;
export type TTripKpi = t.TypeOf<typeof IOTripKpi>;
export type TTripCoop = t.TypeOf<typeof IOTripCoop>;
// ...
export const SavedFileInfo = t.strict({
  fileFormat: ioTypeFromEnum<UsedFileFormatEnum>('UsedFileFormatEnum', UsedFileFormatEnum),
  fileName: t.string,
  fileSize: t.number,
  folder: tt.uuid,
  id: tt.uuid,
});

export type SavedFileInfo = t.TypeOf<typeof SavedFileInfo>;

export const InnerCityTransportRequestInfo = t.strict({
  author: Employee,
  passenger: Employee,
  transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
  desiredDate: t.number,
  expected: RequestRoute,
  segments: t.array(Segment),
  waypoints: t.array(IOWaypoint),
  purpose: TripPurpose,
  compensationType: ioTypeFromEnum<TransportCompensations>('TransportCompensations', TransportCompensations),
  compensationDocuments: t.array(SavedFileInfo),
});

export type InnerCityTransportRequestInfo = t.TypeOf<typeof InnerCityTransportRequestInfo>;

const TripData = t.strict({
  metroTicketsQuantity: t.number,
  tramTicketsQuantity: t.number,
  trolleybusTicketsQuantity: t.number,
  busTicketsQuantity: t.number,
});

export type TripData = t.TypeOf<typeof TripData>;

const tariffData = t.strict({
  metroTicketCost: t.number,
  tramTicketCost: t.number,
  trolleybusTicketCost: t.number,
  busTicketCost: t.number,
  metroAvailability: t.boolean,
  tramAvailability: t.boolean,
  trolleybusAvailability: t.boolean,
  busAvailability: t.boolean,
});

export const CityTripCompensation = t.strict({
  author: Employee,
  passenger: Employee,
  transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
  desiredDate: t.number,
  expected: RequestRoute,
  segments: t.array(Segment),
  waypoints: t.array(IOWaypoint),
  purpose: TripPurpose,
  compensationType: ioTypeFromEnum<TransportCompensations>('TransportCompensations', TransportCompensations),
  compensationDocuments: t.array(SavedFileInfo),

  tariffId: t.string,
  tripData: TripData,
  tariffData,
});

export type CityTripCompensation = t.TypeOf<typeof CityTripCompensation>;

export const TransportCompensation = t.intersection([
  t.type({
    compensationType: AvailablePublicTTCompensationType,
    ticketsCost: tt.money,
  }),
  t.partial({
    transportType: AvailablePublicTransportType,
    ticketsCount: t.number,
    ticketsExpirationStart: t.string,
    ticketsExpirationEnd: t.string,
    attachedDocumentId: t.string,
  }),
]);

export type TransportCompensation = t.TypeOf<typeof TransportCompensation>;

export const PublicTripCompensation = t.intersection([
  t.type({
    author: Employee,
    passenger: Employee,
    purpose: TripPurpose,
    transportCompensation: t.array(TransportCompensation),
  }),
  t.partial({
    expected: RequestRoute,
    segments: t.array(Segment),
    waypoints: t.array(IOWaypoint),
    transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
    desiredDate: t.number,
    compensationDocuments: t.array(SavedFileInfo),
    tariffId: t.string,
    timeZone: t.string,
  }),
]);

export type PublicTripCompensation = t.TypeOf<typeof PublicTripCompensation>;

export const IOCoopTripsSettings = t.type({
  id: t.string,
  transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
  organizationId: t.string,
  economyIndicationYellowRangeLowerBorder: t.number,
  economyIndicationYellowRangeUpperBorder: t.number,
  settings: t.unknown,
});

export type TCoopTripsSettings = t.TypeOf<typeof IOCoopTripsSettings>;